% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mch.R
\name{mch}
\alias{mch}
\title{multicore cluster hybrid}
\usage{
mch(cl, cores, func, params)
}
\description{
#' Note: You need to install multicore in order to use mch:
#' install.packages('multicore',,'http://www.rforge.net/')
}
